% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDyads_or_Triads.R
\name{create.missingness.matrix}
\alias{create.missingness.matrix}
\title{Create family matrix of data missingness}
\usage{
create.missingness.matrix(data.in = stop("No data given!", call. =
  FALSE), new.ids = stop("No IDs given", call. = FALSE))
}
\arguments{
\item{data.in}{The data object (in format as the output of \link{genDataRead}); note
that the design of the data is assumed to be triad.}

\item{new.ids}{A list with pedIndex and ids, as generated by \link{f.check.unique.ids}.}
}
\value{
A matrix where each member of the family has TRUE or FALSE
  (whether they have any genetic data)
}
\description{
Help function that checks the family structure, based on IDs and provided data in .ped
  file, and then creates a matrix with families, where each indiv. has TRUE/FALSE value,
  depending on whether they have or not genetic data.
}
\keyword{internal}
