% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.check.unique.ids.R
\name{f.check.unique.ids}
\alias{f.check.unique.ids}
\title{Checking uniqueness of individuals' IDs}
\usage{
f.check.unique.ids(data.cov)
}
\arguments{
\item{data.cov}{A matrix with "id.fam", "id.c", "id.m", and "id.f" columns.}
}
\value{
List with:
  \itemize{
    \item ids - new IDs in the same order as "id.c" column in the input data
    \item pedIndex - matrix with "family", "id.child", "id.mother", and
       "id.father" columns
  }
}
\description{
This checks whether the individuals have unique IDs. If not, it creates new IDs by
  adding the family ID to the indiv. ID. Since the function checks for specific
  column names, this basically means that the input should be as read from 
  a PED file.
}
\keyword{internal}
