% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.extract.ff.numeric.R
\name{f.extract.ff.numeric}
\alias{f.extract.ff.numeric}
\title{Extract a part of an ff array (matrix) and return it as numeric matrix}
\usage{
f.extract.ff.numeric(gen.data, cols = NULL, rows = NULL)
}
\arguments{
\item{gen.data}{The ff matrix with genetic data.}

\item{cols}{The sequence numbering the columns which are to be retrieved, if NULL (default) then all the columns will be extracted.}

\item{rows}{The sequence numbering the rows which are to be retrieved, if NULL (default) then all the rows will be extracted.}
}
\value{
Matrix with all the requested columns, data in numeric representation.
}
\description{
This function returns the requested columns and/or rows and converts it to a numeric matrix (by default, the extraction from an ff matrix with defined levels returns a factor vector).
}
\keyword{internal}
