% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.get.gen.data.cols.R
\name{f.get.gen.data.cols}
\alias{f.get.gen.data.cols}
\title{Getter for the list with genetic or environmental data}
\usage{
f.get.gen.data.cols(gen.data, cols, by.colname = FALSE)
}
\arguments{
\item{gen.data}{The list with genetic/environmental data (in ff format).}

\item{cols}{The sequence numbering (or listing) the columns which are to be retrieved; this should enumerate the columns as if all the data was in one large matrix.}

\item{by.colname}{Logical: is 'cols' given as a number sequence (FALSE, default) or character vector (TRUE) naming the columns to choose?}
}
\value{
Matrix (in ff format) with all the requested columns.
}
\description{
This function returns the columns requested from the column-wise splitted genetic/environmental data in ff format.
}
\keyword{internal}
