% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finishParallelRun.R
\name{finishParallelRun}
\alias{finishParallelRun}
\title{Closing the Rmpi cluster}
\usage{
finishParallelRun()
}
\description{
This function closes all the slaves spawned in \link{initParallelRun} and
  finishes the mpi routines. This function MUST BE called after all the \link{haplinSlide}
  calls and right before exiting the script/R session!
}
