% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParallelRun.R
\name{initParallelRun}
\alias{initParallelRun}
\title{Initialization of the Rmpi cluster}
\usage{
initParallelRun(cpus)
}
\arguments{
\item{cpus}{Number of cores to use (optional). If given, only that number of CPUs 
will be used. By default (if not set), the Rmpi will check how many CPUs are
available in the system and take the maximum number.}
}
\description{
This function prepares a cluster using \code{Rmpi} package. The initialization
  is paired with closing of the cluster using the \code{\link{finishParallelRun}} function.
}
