% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nfam.R
\name{nfam}
\alias{nfam}
\title{Count the number of families in the data}
\usage{
nfam(data.in)
}
\arguments{
\item{data.in}{The data read in by \link{genDataRead}.}
}
\value{
How many families (integer).
}
\description{
This is a help function to count the number of families in an object read in with
 \link{genDataRead} (or loaded with \link{genDataLoad}). Note: it is assumed that
 the study design is either 'triad' or 'cc.triad'.
}
