\name{print.haplin}
\alias{print.haplin}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print a haplin object}
\description{
 Print basic information about a haplin object
}
\usage{
\method{print}{haplin}(x, ...) 
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{A \code{haplin} object, i.e. the result of running \code{haplin}.}
  \item{...}{Other arguments, passed on to \code{print}.}
}
%\details{}
\references{ Gjessing HK and Lie RT. Case-parent triads: Estimating single- and double-dose effects of fetal and maternal disease gene haplotypes. Annals of Human Genetics (2006) 70, pp. 382-396.\cr\cr
Web Site: \url{http://www.uib.no/smis/gjessing/genetics/software/haplin/}}
\author{Hakon K. Gjessing\cr
Professor of Biostatistics\cr
Division of Epidemiology\cr
Norwegian Institute of Public Health\cr
\email{hakon.gjessing@fhi.no}}
\note{Further information is found on the web page
}
\seealso{\code{\link{haplin}}}






