% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPheno.R
\name{showPheno}
\alias{showPheno}
\title{Display phenotype part of data}
\usage{
showPheno(data.in, n = 5, from, to, sex)
}
\arguments{
\item{data.in}{The data read in by \link{genDataRead}).}

\item{n}{Number of rows to display or "all" (default: 5).}

\item{from}{From which row to display (optional, default: from the first).}

\item{to}{To which row to display (optional).}

\item{sex}{If the sex column is part of the phenotypic information, the user can
choose based on one of the categories used in this column (optional);
NB: this does not combine with the 'to' and 'from' arguments.}
}
\value{
A table with phenotypic and covariate data (if any) extracted from 'data.in'.
}
\description{
This is a help function to extract phenotypic (and covariate) data from an object
read in with \link{genDataRead} (or loaded with \link{genDataLoad}).
}
