% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showSNPnames.R
\name{showSNPnames}
\alias{showSNPnames}
\title{Display marker names}
\usage{
showSNPnames(data.in, n = 5, from, to)
}
\arguments{
\item{data.in}{The data as outputted by \link{genDataRead}, \link{genDataLoad}, 
\link{genDataGetPart} or \link{genDataPreprocess}.}

\item{n}{Number of names to display or "all" (default: 5).}

\item{from}{From which marker to display (optional, default: from the first).}

\item{to}{To which marker to display (optional).}
}
\value{
A vector with marker names, as read in from map.file or generated dummy names.
}
\description{
This is a help function to extract marker names from an object generated by
 \link{genDataRead}, \link{genDataLoad}, \link{genDataGetPart} or 
 \link{genDataPreprocess}.
}
